Enhancement: keystone v2 users endpoint as subservice
- Enhancement: Make the model property part of the base
- Enhancement: Move the users functionality into its own subservice
