+++ openstackinabox/services/keystone/v2/base.py
@@ -1,11 +1,51 @@
+import re
+
+from openstackinabox.models.keystone import KeystoneModel
 from openstackinabox.services.base_service import BaseService
 from openstackinabox.services.keystone.v2.exceptions import *
 
 
 class KeystoneV2ServiceBase(BaseService):
 
+    # USER_ID_PATH_REGEX = re.compile('^\/users\/[a-zA-Z]+[\w\.@-]*$')
+    USER_ID_REGEX = '([0-9]+)'
+    USER_ID_PATH_REGEX = re.compile('^\/users\/{0}$'
+                                    .format(USER_ID_REGEX))
+    USER_ID_KSADM_CREDENTIAL_PATH_REGEX = re.compile(
+        '^\/users\/{0}/OS-KSADM/credentials$'
+        .format(USER_ID_REGEX))
+
+    @staticmethod
+    def get_user_id_from_path(uri_path):
+        uri_matcher = None
+
+        regexes = [
+            KeystoneV2ServiceBase.USER_ID_PATH_REGEX,
+            KeystoneV2ServiceBase.USER_ID_KSADM_CREDENTIAL_PATH_REGEX,
+        ]
+
+        for r in regexes:
+            uri_matcher = r.match(uri_path)
+            if uri_matcher is not None:
+                break
+
+        userid = uri_matcher.groups()[0]
+        return userid
+
     def __init__(self, *args, **kwargs):
         super(KeystoneV2ServiceBase, self).__init__(*args, **kwargs)
+        self.__model = None
+
+    @property
+    def model(self):
+        return self.__model
+
+    @model.setter
+    def model(self, value):
+        if isinstance(value, KeystoneModel):
+            self.__model = value
+        else:
+            raise TypeError('model is not an instance of KeystoneModel')
 
     def helper_validate_token(self, request_headers,
                               enforce_admin, service_admin):
